fix: limit 20chars quotes

+++ server/api/dashboard/customize/set-quotes.ts
@@ -5,6 +5,7 @@ export default defineEventHandler(async (event) => {
     if (!Array.isArray(quotes)) return new Response('Quotes must be an array', {status: 400});
     if (quotes.length === 0) return new Response('Quotes cannot be empty', {status: 400});
     if (quotes.length > 3) return new Response('Quotes cannot be more than 3', {status: 400});
+    if (quotes.some((quote: string) => quote.length > 20)) return new Response('Quotes cannot be more than 20 characters', {status: 400});
 
     await event.context.db.quotes.deleteMany({where: {account_id: event.context.user.id}});
     await event.context.db.quotes.createMany({
